Move TizenApplicationPackageFactory into a separate file
